Move scheduler's SchedulerExtender interface to core/extenders.go
